create individual service methods
