+++ src/pages/Buckets/Container.tsx
@@ -2,9 +2,9 @@ import { useEffect } from "react"
 import { useDispatch, useSelector } from "react-redux"
 import { selectAddress } from "../../redux/slices/authenticationSlice"
 import {
-	getAllBuckets,
 	selectBuckets,
 	selectBucketsLoading,
+	userGetAllBuckets,
 } from "../../redux/slices/bucketSlice"
 import { AppDispatch } from "../../redux/store"
 import BucketsPageComponent from "./Component"
@@ -15,7 +15,7 @@ function BucketsPageContainer() {
 	const loading = useSelector(selectBucketsLoading)
 	const address = useSelector(selectAddress)
 	useEffect(() => {
-		dispatch(getAllBuckets(address))
+		dispatch(userGetAllBuckets(address))
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [])
