+++ src/pages/Schemas/Container.tsx
@@ -8,9 +8,8 @@ import {
 	userGetAllSchemas,
 } from "../../redux/slices/schemasSlice"
 import { AppDispatch } from "../../redux/store"
-import { SchemaMeta } from "../../service/schemas"
 import { MODAL_CONTENT_NEW_SCHEMA } from "../../utils/constants"
-import { IsearchableList, IsearchableListItem } from "../../utils/types"
+import { IsearchableList, IsearchableListItem, SchemaMeta } from "../../utils/types"
 import SchemasPageComponent from "./Component"
 import EmptyList from "./components/EmptyList"
 import ViewProperties from "./components/ViewProperties"
