+++ src/utils/checkLogin.ts
@@ -1,4 +1,4 @@
-import { getAccountInfo } from "../service/account"
+import getAccountInfo from "../service/getAccountInfo"
 import { getAppStateFromLocalCache } from "./localStorage"
 
 export async function isLoggedInServer() {
